+++ changelog.md
@@ -1,3 +1,12 @@
+# 1.2.0
+### Additions
+- Added Russian translations (thank you CreaterIDK634!)
+
+### Changes
+- Updated for Farmer's Delight 1.3
+- Noodle Soup and Baked Cod Stew now support boiled & sliced egg ingredients
+- Improve tooltips for Dyed Egg
+
 # 1.1.0
 ### Additions
 - Dyed eggs are now placeable!